Extracted successful match logic in own class
