+++ imports/api/games/server/matchMaking/ImmediateMatchMaker.js
@@ -1,6 +1,6 @@
 import {MatchMakers} from '/imports/api/games/matchMakers.js';
-import GameCreator from '/imports/api/games/server/GameCreator.js';
 import MatchMaker from '/imports/api/games/server/matchMaking/MatchMaker.js';
+import UserMatch from '/imports/api/games/server/matchMaking/UserMatch.js';
 import {Meteor} from "meteor/meteor";
 import {Random} from 'meteor/random';
 
@@ -20,41 +20,7 @@ export default class ImmediateMatchMaker extends MatchMaker {
 				match = MatchMakers.findOne({modeSelection: modeSelection, tournamentId: tournamentId});
 				const matchedUsers = this.matchedUsers(match);
 				if (matchedUsers.length > 0) {
-					match = MatchMakers.findOne({modeSelection: modeSelection, tournamentId: tournamentId});
-					const gameId = GameCreator.fromMatchMaker(matchedUsers, modeSelection, tournamentId);
-
-					//Remove from usersToMatch
-					let remainingUsersToMatch = match.usersToMatch;
-					for (let userId of matchedUsers) {
-						const index = remainingUsersToMatch.indexOf(userId);
-						if (index !== -1) {
-							remainingUsersToMatch.splice(index, 1);
-						}
-					}
-					MatchMakers.update({_id: match._id}, {$set: {usersToMatch: remainingUsersToMatch}});
-
-					//Add to matched with gameId
-					let matched = match.matched.concat({
-						users: matchedUsers,
-						gameId: gameId
-					});
-					MatchMakers.update({_id: match._id}, {$set: {matched: matched}});
-
-					//Remove matched after a while
-					Meteor.setTimeout(() => {
-						match = MatchMakers.findOne({modeSelection: modeSelection, tournamentId: tournamentId});
-
-						matched = match.matched;
-						for (let i = 0; i < matched.length; i++) {
-							let item = matched[i];
-							if (matched[i].gameId === gameId) {
-								matched.splice(i, 1);
-								break;
-							}
-						}
-
-						MatchMakers.update({_id: match._id}, {$set: {matched: matched}});
-					}, 10000);
+					UserMatch.match(modeSelection, tournamentId, matchedUsers);
 				}
 			}
 		} else {
